refactor(blockchain-link): base.ts -> baseWorker.ts
